Added a new grouped barchart function!
With the new function make_grouped_plots it is possible to automatically create grouped bargraphs for related samples. The function tries to estimate which samples belong together by reading their names and finding similarity. It can then either generate one grouped plot for each grouping or an assembled figure with subplots in it.