Added Filters
AND The entire architecture of how results are stored was changed and now directly and always works with pandas dataframes.

This was necessary to preserve index information about which replicates were excluded by filters.

Now (to my knowledge and tests) everything works like a charm, both for manual and pipelines.